Fix TRELLIS.2 on ROCm gfx1201
Chunk large F.linear/GEMMs, compare occupancy in fp32, prefer bf16 VAEs,
and keep LSCM/remesh/GLB export off broken hipBLAS and bfloat16 numpy paths.

+++ comfy_extras/nodes_save_3d.py
@@ -139,11 +139,11 @@ def save_glb(vertices, faces, filepath=None, metadata=None,
     emissive_image: PIL.Image - Optional emissive (glow) texture, written as emissiveTexture.
     """
 
-    # Convert tensors to numpy arrays
-    vertices_np = vertices.cpu().numpy().astype(np.float32)
+    # Convert tensors to numpy arrays (float() so bf16 vertex colors survive .numpy())
+    vertices_np = vertices.float().cpu().numpy()
     faces_signed = faces.cpu().numpy().astype(np.int64)
-    uvs_np = uvs.cpu().numpy().astype(np.float32) if uvs is not None else None
-    colors_np = vertex_colors.cpu().numpy().astype(np.float32) if vertex_colors is not None else None
+    uvs_np = uvs.float().cpu().numpy() if uvs is not None else None
+    colors_np = vertex_colors.float().cpu().numpy() if vertex_colors is not None else None
     if colors_np is not None:
         colors_np = np.clip(colors_np, 0.0, 1.0)
 
@@ -166,12 +166,12 @@ def save_glb(vertices, faces, filepath=None, metadata=None,
             f"save_glb: vertex_colors has {colors_np.shape[0]} entries but vertex count is {n_verts}"
         )
 
-    normals_np = normals.cpu().numpy().astype(np.float32) if normals is not None else None
+    normals_np = normals.float().cpu().numpy() if normals is not None else None
     if normals_np is not None and normals_np.shape[0] != n_verts:
         raise ValueError(
             f"save_glb: normals has {normals_np.shape[0]} entries but vertex count is {n_verts}"
         )
-    tangents_np = tangents.cpu().numpy().astype(np.float32) if tangents is not None else None
+    tangents_np = tangents.float().cpu().numpy() if tangents is not None else None
     if tangents_np is not None and tangents_np.shape != (n_verts, 4):
         raise ValueError(
             f"save_glb: tangents must be (N, 4) with N={n_verts}, got {tuple(tangents_np.shape)}"
@@ -505,7 +505,7 @@ def mesh_item_to_glb_bytes(mesh, index, metadata=None):
         t = getattr(mesh, attr, None)
         if t is None:
             return None
-        a = (t[index].clamp(0.0, 1.0).cpu().numpy() * 255).astype(np.uint8)
+        a = (t[index].clamp(0.0, 1.0).float().cpu().numpy() * 255).astype(np.uint8)
         assert a.ndim == 3 and a.shape[-1] == 3, f"{attr} must be (B, H, W, 3), got {tuple(t.shape)}"
         return Image.fromarray(a, mode="RGB")
